fix(audit): pass 7 — close out deferred items per skill contract
The /code-audit skill explicitly forbids deferred items. Pass-6's
"deferred" list was a contract violation. This pass closes every
item with the appropriate root-cause fix.

Items closed:

1. overlay.cpp file-size cap (was 893 / over 800):
   Extracted rekeyOverlayState + validateScreenStateInvariant +
   installOverlayGeometryWatcher into new overlayservice/rekey.cpp
   (244 lines). overlay.cpp drops to 702 lines, under the cap. The
   three functions form a cohesive cluster: rekey invokes the
   geometry watcher post-move; the debug-only invariant check
   verifies cross-side pointer alignment that rekey is the primary
   risk-source for. src/CMakeLists.txt updated.

2. Em-dashes in code comments (project rule: "no AI-ism punctuation
   in user prose ... comments"):
   Mechanical sweep across all 36 PR-scope files. U+2014 (em-dash)
   and U+2013 (en-dash) replaced with ASCII hyphen-minus. Affects
   comment text only — no identifier or executable-code changes.

3. Daemon-bridge unit-test coverage:
   Added test_overlay_smoke case `hideSlotFiresCompletionWhenSlotItemIsNullQPointer`
   pinning the contract the daemon's PerScreenOverlayState slot
   accessors rely on (SlotEntry exists in map but its QPointer<QQuickItem>
   is null because the underlying item was destroyed mid-lifecycle).
   Total smoke test cases: 22 (was 21).

4. <QChar> include in shellhost.cpp:
   False alarm from pass-5 deferred list. QLatin1Char (used in
   makePerInstanceRole) is defined in <qchar.h>, so the include is
   load-bearing. No change.

5. validateScreenStateInvariant cross-side pointer-alignment check:
   Added a debug-only assertion that for every key in m_screenStates
   with a non-null `shell` borrowed pointer, the daemon's pointer
   matches the lib's stateFor(key) view. Catches desync from a
   future rekey/ensureShell drift or out-of-band lib mutation.
   Uses an explicit `const PhosphorOverlay::ShellHost*` binding to
   reach the const stateFor overload (returns pointer-or-nullptr;
   the non-const overload would materialize an entry on a miss,
   which a debug-only check must not do).

Build clean. 178/178 ctest pass (+1 new smoke case = 22 within the
overlay-lib smoke executable, ctest counts the executable as 1).
