+++ libs/phosphor-overlay/include/PhosphorOverlay/SlotEntry.h
@@ -9,7 +9,7 @@
 // SurfaceAnimator show/hide leg targets.
 //
 // The role is per-slot because every consumer's slots have a stable
-// animator scope — PZ's OSD slot always animates as PzRoles::Osd, the
+// animator scope - PZ's OSD slot always animates as PzRoles::Osd, the
 // zone-selector slot as PzRoles::ZoneSelector, and so on. Pinning the
 // role into the slot entry lets the library drive
 // SurfaceAnimator::beginShow / beginHide without the consumer
@@ -27,7 +27,7 @@ namespace PhosphorOverlay {
 
 struct PHOSPHOROVERLAY_EXPORT SlotEntry
 {
-    /// The slot's QQuickItem. Borrowed — owned by the shell window's
+    /// The slot's QQuickItem. Borrowed - owned by the shell window's
     /// scene graph; QPointer auto-clears if the underlying Item is
     /// destroyed (typically: shell torn down).
     QPointer<QQuickItem> item;
