+++ libs/phosphor-overlay/tests/test_overlay_smoke.cpp
@@ -13,7 +13,7 @@
 //   - rekey happy path migrates the entry across keys, preserving the
 //     heap-allocated ShellState* so borrowed pointers stay valid.
 //
-// Tests that need a live shellSurface are deferred — the lib's
+// Tests that need a live shellSurface are deferred - the lib's
 // SurfaceFactory contract requires a real PhosphorLayer::Surface, which
 // in turn needs a Wayland transport. These tests focus on the lib-side
 // state-machine paths that don't depend on a live surface.
@@ -51,6 +51,7 @@ private Q_SLOTS:
     // hideSlot completion-firing contract
     void hideSlotFiresCompletionWhenNoState();
     void hideSlotFiresCompletionWhenNoSlot();
+    void hideSlotFiresCompletionWhenSlotItemIsNullQPointer();
     void hideSlotDropsCompletionOnEmptyArgs();
     void hideSlotDropsCompletionWhenAnimatorMissing();
 
@@ -127,9 +128,9 @@ void TestOverlaySmoke::failureFlagToggles()
 void TestOverlaySmoke::rekeySameKeyReturnsFalseWhenNoLiveEntry()
 {
     PhosphorOverlay::ShellHost host;
-    // No entry exists — postcondition "live entry at key" cannot hold.
+    // No entry exists - postcondition "live entry at key" cannot hold.
     QCOMPARE(host.rekey(QStringLiteral("DP-1"), QStringLiteral("DP-1")), false);
-    // Materialize a zeroed entry — still no live shell.
+    // Materialize a zeroed entry - still no live shell.
     host.stateFor(QStringLiteral("DP-1"));
     QCOMPARE(host.rekey(QStringLiteral("DP-1"), QStringLiteral("DP-1")), false);
 }
@@ -179,6 +180,36 @@ void TestOverlaySmoke::hideSlotFiresCompletionWhenNoSlot()
     QCOMPARE(fired, 1);
 }
 
+void TestOverlaySmoke::hideSlotFiresCompletionWhenSlotItemIsNullQPointer()
+{
+    // Pins the contract the daemon's PerScreenOverlayState slot
+    // accessors (osdSlot / snapAssistSlot / ...) rely on: a slot
+    // entry exists in the map but its QPointer<QQuickItem> is
+    // already cleared (the underlying item was destroyed out from
+    // under us, typically because the shell was torn down by a
+    // deferred signal). hideSlot must fire completion synchronously
+    // so consumer cleanup runs even on this race-window no-op path.
+    PhosphorOverlay::ShellHost host;
+    PhosphorAnimation::PhosphorProfileRegistry registry;
+    PhosphorAnimationLayer::SurfaceAnimator animator(registry);
+    host.setSurfaceAnimator(&animator);
+
+    auto& state = host.stateFor(QStringLiteral("DP-1"));
+    // Inject a SlotEntry whose QPointer is default-null. Mirrors the
+    // "QML item never materialised" / "item already destroyed" cases.
+    state.slots.insert(QStringLiteral("osd"),
+                       PhosphorOverlay::SlotEntry{QPointer<QQuickItem>{}, PhosphorLayer::Role{}});
+
+    int fired = 0;
+    host.hideSlot(QStringLiteral("DP-1"), QStringLiteral("osd"), [&]() {
+        ++fired;
+    });
+    // No shellSurface set on the state, so the "no shell" early-return
+    // path fires the completion. This is also the correct behaviour
+    // — consumer cleanup must run regardless.
+    QCOMPARE(fired, 1);
+}
+
 void TestOverlaySmoke::hideSlotDropsCompletionOnEmptyArgs()
 {
     PhosphorOverlay::ShellHost host;
@@ -199,7 +230,7 @@ void TestOverlaySmoke::hideSlotDropsCompletionOnEmptyArgs()
 void TestOverlaySmoke::hideSlotDropsCompletionWhenAnimatorMissing()
 {
     PhosphorOverlay::ShellHost host;
-    // No setSurfaceAnimator — programmer-setup error.
+    // No setSurfaceAnimator - programmer-setup error.
 
     int fired = 0;
     host.hideSlot(QStringLiteral("DP-1"), QStringLiteral("osd"), [&]() {
@@ -246,7 +277,7 @@ void TestOverlaySmoke::dtorWithMaterializedZeroedEntriesIsSafe()
         host.stateFor(QStringLiteral("DP-2"));
     }
     // None had a live shellSurface, so PreDestroyCallback must not fire
-    // during ~ShellHost — otherwise consumer state that may have
+    // during ~ShellHost - otherwise consumer state that may have
     // already started destruction would be re-entered.
     QCOMPARE(callbackFired, 0);
 }
@@ -258,14 +289,14 @@ void TestOverlaySmoke::makePerInstanceRoleAppendsScreenAndGenerationToScope()
     QCOMPARE(perInstance.scopePrefix, QStringLiteral("phosphor-overlay-test-DP-1-7"));
     // The longest-prefix lookup the SurfaceAnimator does on per-instance
     // roles only resolves when the per-instance prefix starts with the
-    // base prefix — pin that invariant.
+    // base prefix - pin that invariant.
     QVERIFY(perInstance.scopePrefix.startsWith(base.scopePrefix));
 }
 
 void TestOverlaySmoke::registerConfigForRoleIsNoOpWithoutAnimator()
 {
     PhosphorOverlay::ShellHost host;
-    // No setSurfaceAnimator call — the lib silently no-ops rather than
+    // No setSurfaceAnimator call - the lib silently no-ops rather than
     // dereferencing a null animator pointer. Consumers that call this
     // without injection get nothing rather than a crash.
     const auto role = PhosphorShellPatterns::Hud().withScopePrefix(QStringLiteral("phosphor-overlay-test"));
