+++ libs/phosphor-zones/include/PhosphorZones/IZoneLayoutRegistry.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-// IZoneLayoutRegistry — enumeration + mutation of the catalog of
+// IZoneLayoutRegistry - enumeration + mutation of the catalog of
 // manual zone layouts.
 //
 // Split out of ILayoutManager so callers that need the layout set
@@ -14,7 +14,7 @@
 //
 // Inherits PhosphorLayout::ILayoutSourceRegistry so concrete registries
 // (LayoutManager) carry the unified `contentsChanged` signal that
-// ZonesLayoutSource subscribes to — matching the pattern every other
+// ZonesLayoutSource subscribes to - matching the pattern every other
 // provider library (phosphor-tiles, future phosphor-scrolling, …)
 // follows. Inheriting QObject via the unified base rather than
 // directly keeps ILayoutManager's non-virtual multi-inheritance safe:
@@ -53,7 +53,7 @@ public:
     explicit IZoneLayoutRegistry(QObject* parent = nullptr);
     ~IZoneLayoutRegistry() override;
 
-    /// Enumerate every known layout. Borrowed pointers — owned by the
+    /// Enumerate every known layout. Borrowed pointers - owned by the
     /// concrete registry (typically @c LayoutManager). Order is the
     /// registry's natural iteration order.
     virtual QVector<Layout*> layouts() const = 0;
@@ -66,23 +66,23 @@ public:
     /// layout with that id is known to the registry.
     virtual Layout* layoutById(const QUuid& id) const = 0;
 
-    /// @param layout Ownership transferred — the registry adopts @p layout
+    /// @param layout Ownership transferred - the registry adopts @p layout
     ///               and is responsible for its lifetime from this call on.
     virtual void addLayout(Layout* layout) = 0;
-    /// @param layout Borrowed — caller hands the pointer in; the
+    /// @param layout Borrowed - caller hands the pointer in; the
     ///               registry un-registers it and schedules deletion
     ///               via @c deleteLater (matching how the registry
     ///               adopted it in @ref addLayout). Callers must drop
     ///               any other references before this call returns.
     virtual void removeLayout(Layout* layout) = 0;
     virtual void removeLayoutById(const QUuid& id) = 0;
-    /// @param source Borrowed — caller retains ownership.
+    /// @param source Borrowed - caller retains ownership.
     /// @return       Newly allocated copy; ownership transferred to the
     ///               registry (mirrors @c addLayout semantics). Returns
     ///               nullptr if @p source is unknown.
     virtual Layout* duplicateLayout(Layout* source) = 0;
 
-    // Active layout (internal — used for resnap / geometry / overlay
+    // Active layout (internal - used for resnap / geometry / overlay
     // machinery). Borrowed pointer owned by the registry.
     virtual Layout* activeLayout() const = 0;
     virtual void setActiveLayout(Layout* layout) = 0;
@@ -93,7 +93,7 @@ public:
     // These queries resolve a layout for a (screen, desktop, activity)
     // context by walking the assignment cascade and falling back to the
     // global default. Overlay/geometry/animation consumers depend on this
-    // shape, so it lives on the interface — callers can target the
+    // shape, so it lives on the interface - callers can target the
     // contract without depending on the concrete @c LayoutRegistry.
 
     /// Cascade-resolve the manual layout for @p screenId. Returns
@@ -131,14 +131,14 @@ Q_SIGNALS:
 
     // Active-layout selection. Fires from @c setActiveLayout /
     // @c setActiveLayoutById only when the active-layout pointer actually
-    // changes (concrete implementer guards with an equality check —
+    // changes (concrete implementer guards with an equality check -
     // matches the project rule "only emit signals when value actually
     // changes").
     void activeLayoutChanged(Layout* layout);
 
     // Assignment churn. Fires when a (screenId, virtualDesktop)
     // assignment changes; activity context is intentionally omitted
-    // from the signal — consumers that care about activity-keyed
+    // from the signal - consumers that care about activity-keyed
     // assignments re-query via @c layoutForScreen with their current
     // activity. Concrete impl emits only on actual change (matches the
     // project "emit only when value changes" rule).
