+++ src/core/unifiedlayoutlist.h
@@ -5,13 +5,13 @@
 
 // Picker-composition helpers for the unified layout list.
 //
-// A "unified layout" is one row in the layout-picker UI — it can back either
+// A "unified layout" is one row in the layout-picker UI - it can back either
 // a manual zone-based layout OR an autotile algorithm, and the composition
 // functions stitch both sources into a single sorted list for the overlay /
 // zone selector / D-Bus layout list.
 //
 // The canonical entry type is @c PhosphorLayout::LayoutPreview (from
-// phosphor-layout-api). All helpers in this header operate on it directly —
+// phosphor-layout-api). All helpers in this header operate on it directly -
 // there is no separate app-layer mirror struct.
 //
 // Lives OUTSIDE libs/ because it pulls in autotile's PhosphorTiles::
@@ -60,12 +60,12 @@ using ::PlasmaZones::IOrderingSettings;
  *
  * When @p includeAutotile is true the helper needs a way to enumerate
  * autotile previews. It picks the input as follows:
- *   1. @p autotileSource — a long-lived @c PhosphorLayout::ILayoutSource
+ *   1. @p autotileSource - a long-lived @c PhosphorLayout::ILayoutSource
  *      (typically the autotile source owned by a
  *      @c PhosphorLayout::LayoutSourceBundle) whose internal preview cache
  *      is reused across calls. Pass the composition root's bundle source
- *      here when available — this is the fast path.
- *   2. @p algorithmRegistry — fallback. When @p autotileSource is null
+ *      here when available - this is the fast path.
+ *   2. @p algorithmRegistry - fallback. When @p autotileSource is null
  *      the helper constructs a transient @c AutotileLayoutSource over the
  *      registry for this one call. Cache is discarded between calls.
  * Either must be non-null when @p includeAutotile is true; the registry
@@ -91,7 +91,7 @@ buildUnifiedLayoutList(PhosphorZones::IZoneLayoutRegistry* layoutManager,
  * indicates whether the layout matches the current screen's aspect ratio.
  *
  * See the non-filtered overload for @p autotileSource / @p algorithmRegistry
- * semantics — same fallback rules apply.
+ * semantics - same fallback rules apply.
  */
 PLASMAZONES_EXPORT QVector<PhosphorLayout::LayoutPreview>
 buildUnifiedLayoutList(PhosphorZones::IZoneLayoutRegistry* layoutManager,
